configs: organize configs
