+++ .github/workflows/autotest/gem5-h.cfg
@@ -102,7 +102,7 @@ pre-task =
 task = {set_var} && \
     ./build/RISCV/gem5.opt \
     --outdir=$sublog$ \
-    ./configs/example/xiangshan.py \
+    ./configs/example/kmhv2.py \
     --enable-h-gcpt \
     --generic-rv-cpt=$binfile$ 
 post-task = echo run successfully at $binfile$
