+++ .github/workflows/autotest/gem5-vec.cfg
@@ -39,7 +39,7 @@ except-task =
 binpath = /nfs/home/share/gem5_ci/vector_test/*.bin
 numacores = 0
 pre-task = 
-fs_path = ./configs/example/kmh.py --enable-riscv-vector --restore-rvv-cpt
+fs_path = ./configs/example/kmhv2.py --enable-riscv-vector --restore-rvv-cpt
 task ={set_env} && ./build/RISCV/gem5.opt \
     --outdir=$sublog$ \
     {fs_path} \
