+++ util/xs_scripts/kmh_6wide_h.sh
@@ -7,4 +7,4 @@ for var in GCBH_REF_SO GCBH_RESTORER gem5_home; do
     checkForVariable $var
 done
 
-$gem5 $gem5_home/configs/example/xiangshan.py --generic-rv-cpt=$1 --restore-rvh-cpt
+$gem5 $gem5_home/configs/example/kmhv2.py --generic-rv-cpt=$1 --restore-rvh-cpt
