+++ util/xs_scripts/kmh_6wide_ideal_l2.sh
@@ -7,4 +7,4 @@ for var in GCB_REF_SO GCB_RESTORER gem5_home; do
     checkForVariable $var
 done
 
-$gem5 $gem5_home/configs/example/xiangshan.py --generic-rv-cpt=$1 --mem-type=SimpleMemory --ideal-cache --no-l3cache
+$gem5 $gem5_home/configs/example/kmhv2.py --generic-rv-cpt=$1 --mem-type=SimpleMemory --ideal-cache --no-l3cache
